+++ LookupAnything/Framework/Constants/Constant.cs
@@ -50,13 +50,6 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Constants
             public static string Heater = "Heater";
         }
 
-        /// <summary>The names of locations referenced by the mod.</summary>
-        public static class LocationNames
-        {
-            /// <summary>The internal name for the greenhouse.</summary>
-            public static string Greenhouse = "Greenhouse";
-        }
-
         /// <summary>The names of buildings referenced by the mod.</summary>
         public static class BuildingNames
         {