+++ LookupAnything/release-notes.md
@@ -9,6 +9,7 @@
   * new friendship data;
   * crab pot bait.
 * Added number of item needed in bundle list. (Thanks to StefanOssendorf!)
+* Added support for custom greenhouse locations.
 * Fixed issue where a bundle that needs two stacks of an item won't be listed on the item lookup if one stack is filled. (Thanks to StefanOssendorf!)
 * Fixed Custom Farming Redux machines not drawn correctly in some cases.
 * Improved translations. Thanks to alca259 (Spanish), fadedDexofan (Russian), and TaelFayre (Portuguese)!